cairo: simplify circuit primitives — F-C-2 dead code + F-C-12 auth_idx narrowing
Two pure simplifications in the circuit's trusted base, both purely
making the audit surface smaller without changing public-output
semantics. Other circuit findings (F-C-3 / F-C-4 / F-C-7 / F-C-9 /
F-C-10) are explicitly closed as WONTFIX with the audit-quality
reasoning recorded in findings.md.

F-C-2 — delete dead WOTS+ IV. `blake_hash.cairo` carried a
`blake2s_iv_wots()` ("wotsSP__") and a `pub fn hash1_wots(a)` that
used it, plus a unit test pinning that they had a distinct domain
from `hash1`. Nothing in production code called either. The actual
WOTS+ chain hash uses `hash3_generic` with TAG_XMSS_CHAIN inside the
ADRS, which is RFC-8391-style and sound. The risk was code-quality
only — a future refactor could plausibly switch from one to the
other and silently change the protocol's hash.

  Removed: `blake2s_iv_wots`, `hash1_wots`,
  `test_hash1_wots_uses_distinct_domain`, and the `hash1_wots` import
  in the test module. "The WOTS+ chain hash" is now exactly one
  thing in the codebase.

F-C-12 — narrow `auth_idx` from `u64` to `u32` across all three
circuits. Both `shield::verify` (single `auth_idx: u64`) and
`transfer::verify` / `unshield::verify` (`auth_index_list:
Span<u64>`) were taking u64 and immediately doing
`(*x).try_into().unwrap()` to narrow to u32 for the WOTS+ verify.
The double-narrowing (felt252 -> u64 -> u32) made an auditor
mentally trace the casts to confirm the bound. With `AUTH_DEPTH =
16` the actual constraint is `auth_idx < 65536`, which fits in u16
let alone u32.

  All three `verify` functions now take `u32` directly. Inner
  `try_into().unwrap()` calls and the `auth_idx_u32` rebind in
  shield are gone. The `run_*.cairo` wrappers cast felt252 -> u32
  once at the boundary. Test fixtures dropped the
  `u32 -> u64 -> u32` `.into()` round-trips.

  Internal bytecode change only — no public-output layout change,
  so the verified-bridge fixture's pinned program_hashes still
  match the proofs in the JSON (proofs are pinned to whatever
  bytecode they were generated against; on-disk executables move
  independently). All 13 bridge_flow integration tests still pass
  against the regenerated real-proof fixture without regenerating.

WONTFIX'd in findings.md with explicit rationale (so the next
audit pass doesn't re-raise them):

  F-C-1: 251-bit felt forced; security still vastly above 128-bit.
  F-C-3: length asserts belong in callers, not primitives.
  F-C-4: `try_into().unwrap()` IS the bound check; subsumed by F-C-12.
  F-C-7: kernel-layer policy, not circuit-layer math.
  F-C-9: type system enforces u32 bounds; runtime assert dead code.
  F-C-10: structural invariant, not runtime property.
  F-C-11: 251-bit felt forced (same root cause as F-C-1).

F-C-6 stays open because it's subsumed by F-X-1 (kernel-side
producer-fee owner_tag enforcement).

Workspace 410 tests pass; OCaml 166 pass; Cairo 81 pass (down 1 from
the deleted dead-code test).

+++ cairo/src/run_shield.cairo
@@ -46,7 +46,7 @@ fn main(args: Array<felt252>) -> Array<felt252> {
     let d_j = *args.at(12);
     let rseed = *args.at(13);
     let blind = *args.at(14);
-    let auth_idx: u64 = (*args.at(15)).try_into().unwrap();
+    let auth_idx: u32 = (*args.at(15)).try_into().unwrap();
 
     let wots_start: u32 = fixed_prefix;
     let wots_sig = args.span().slice(wots_start, wots_chains);
